backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pymongo
dnspython
djongo==1.3.7
redis==5.0.1
channels==4.1.0
channels-redis==4.2.0
boto3==1.34.89
django-storages[aws]==1.14.2
Pillow==10.1.0
requests==2.31.0
python-decouple==3.8
pytest==8.3.1
pytest-django==4.9.0
factory-boy==3.3.3
celery==5.4.0
django-extensions==3.2.3
drf-spectacular==0.27.2  # Swagger/OpenAPI
twilio==9.0.0
sendgrid==6.10.0
openai==1.3.6  # AI features
gunicorn==22.0.0
sentry-sdk==2.3.1
django-filter==21.1
django-filter==21.1

